requirements.txt
django-postgrespool==0.3.0
gunicorn==19.6.0
psycopg2==2.6
SQLAlchemy==1.0.4

-r requirements-base.txt